requirements.txt
pandas
numpy
more-itertools
xlrd
openpyxl
Pillow
pybase64
streamlit
plotly-express